feat: Session storage preserves and utilizes the user's location information until they exit the browser
