+++ app/map/page.tsx
@@ -1,4 +1,4 @@
-import MapPage from "@/src/pages/map"
+import MapPage from "@/src/pages/map/page"
 
 export default function Map() {
   return <MapPage />
