+++ app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from "@/src/pages/home"
+import HomePage from "@/src/pages/home/page"
 
 export default function Home() {
   return <HomePage />
