+++ src/utils/constants/sdk.ts
@@ -1,0 +1,1 @@
+export const KAKAO_SDK_URL = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_KAKAO_APP_JS_KEY}&autoload=false`
